Struct probe_c_api::CompileRunOutput

pub struct CompileRunOutput {
    pub compile_output: Output,
    pub run_output: Option<Output>,
}

Outputs of both compilation and running.

Fields

compile_output

Output of the compilation phase.

run_output

Output of the run phase. It is optional because if the compilation failed, we won't try to run at all.

Methods

impl CompileRunOutput

fn successful_run_output(&self) -> CProbeResult<String>

Returns a probe program's standard output as a UTF-8 string.

This function does not panic. If the compilation or run failed, this is reported in the error. If the program's output is not valid UTF-8, lossy conversion is performed.